% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affs.R
\name{affs}
\alias{affs}
\title{affs}
\usage{
affs(x)
}
\arguments{
\item{x}{gene expression matrix where the rows represent cells, and the columns represents the features.}
}
\value{
affinity matrix.
}
\description{
Local scaling affinity.
}
\examples{
x<-readRDS("data/pollen.rds")
data<-logcounts(x)
A<-affs(t(data))
}
