% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self.diffusion.R
\name{self.diffusion}
\alias{self.diffusion}
\title{self.diffusion}
\usage{
self.diffusion(A, K)
}
\arguments{
\item{A}{affinity matrix.}

\item{K}{iteration steps.}
}
\value{
diffusion graph.
}
\description{
Self-diffusion on local scaling affinity matrix.
}
\examples{
x<-readRDS("data/pollen.rds")
data<-logcounts(x)
A<-affs(t(data))
Sda<-self.diffusion(A,4)
}
\keyword{A}
\keyword{K}
