% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.clu.R
\name{spec.clu}
\alias{spec.clu}
\title{spec.clu}
\usage{
spec.clu(affinity, K, type = 3)
}
\arguments{
\item{affinity}{input graph}

\item{K}{the number of clusters}

\item{type}{indicators of variants of spectral clustering}
}
\value{
label
}
\description{
Spectral clustering
}
\examples{
x<-readRDS("data/pollen.rds")
data<-logcounts(x)
A<-affs(t(data))
Sda<-self.diffusion(A,4)
label<-spec.clu(Sda,11)
}
\keyword{K}
\keyword{affinity}
